- move reference counters code
