+++ server/Makefile.am
@@ -11,7 +11,8 @@ snapperd_SOURCES =					\
 	Client.cc		Client.h		\
 	MetaSnapper.cc		MetaSnapper.h		\
 	Background.cc		Background.h		\
-	Types.cc		Types.h
+	Types.cc		Types.h			\
+	RefCounter.cc		RefCounter.h
 
 snapperd_LDADD = ../snapper/libsnapper.la ../dbus/libdbus.la -lrt
