+++ server/MetaSnapper.cc
@@ -34,65 +34,6 @@
 MetaSnappers meta_snappers;
 
 
-RefCounter::RefCounter()
-    : counter(0), last_used(steady_clock::now())
-{
-}
-
-
-int
-RefCounter::inc_use_count()
-{
-    boost::lock_guard<boost::mutex> lock(mutex);
-
-    return ++counter;
-}
-
-
-int
-RefCounter::dec_use_count()
-{
-    boost::lock_guard<boost::mutex> lock(mutex);
-
-    assert(counter > 0);
-
-    if (--counter == 0)
-	last_used = steady_clock::now();
-
-    return counter;
-}
-
-
-void
-RefCounter::update_use_time()
-{
-    boost::lock_guard<boost::mutex> lock(mutex);
-
-    last_used = steady_clock::now();
-}
-
-
-int
-RefCounter::use_count() const
-{
-    boost::lock_guard<boost::mutex> lock(mutex);
-
-    return counter;
-}
-
-
-milliseconds
-RefCounter::unused_for() const
-{
-    boost::lock_guard<boost::mutex> lock(mutex);
-
-    if (counter != 0)
-	return milliseconds(0);
-
-    return duration_cast<milliseconds>(steady_clock::now() - last_used);
-}
-
-
 MetaSnapper::MetaSnapper(ConfigInfo& config_info)
     : config_info(config_info), snapper(NULL)
 {
